Make the `featureLevel` not optional in the user's profile (OSIO#1937)
Make the field in the `user` type as a `string` instead of `*string`.
Make the DB column `NOT NULL` in the DB with a default value to `released`,
and update all existing records.

Also, refactored some code to reduce the amount of lines to init user/indentity.

Fixes openshiftio/openshift.io#1937

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>