Capacity flag per cluster to rate limit tenant resources for che and jenkins (#477)
Part of https://github.com/openshiftio/openshift.io/issues/3320

To avoid overloading clusters we need a way to limit total amount of che and jenkins instances running on each cluster. This PR will add a new parameter to the cluster config which will reflect the current capacity of each cluster. When the corresponding secret is modified it will cause the config file update. The file watcher will re-load cluster configuration.