List spaces for which current user has a role in (#545)
adding a new endpoint in the `/user` base: `/spaces` to list spaces in which the current
user (identitfied by the request token) has a role.
includes changes in the `UserService` and `RoleRepository` to retrieve resources of
a given type for which a user as a direct role, or a role via a team, or
a role via a default role mapping while having a direct role or via a team
on another resource. For example, a user has `admin` role on an org and is
given the `contributor` on the space via a default role mapping.
Custom role mappings are not supported yet.

also:
- refactor `identityIDFromWrapper` as standalone function
- fix the missing call to the gorm callback when adding
a member to a team, by calling the repository instead of
performing a direct SQL insert.

Fixes #545

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>