Remove role names and scopes from API
Roles and scopes should not be provided in the response
to the endpoint that lists all resources of a given type
for which the current user has a role, as this may lead
to bad usage of the API. Other endpoints already exists
to check the permissions on a given resource.

Response now looks like this:
````
{
    "data": [{
        "id": "114c89ec-8ff2-4c65-adb4-161c9b505be9",
        "links": {
            "related": "http:///api/resource/114c89ec-8ff2-4c65-adb4-161c9b505be9"
        },
        "type": "resources"
    }, {
        "id": "5d7d3f9b-5a53-4d7e-b321-de3999542626",
        "links": {
            "related": "http:///api/resource/5d7d3f9b-5a53-4d7e-b321-de3999542626"
        },
        "type": "resources"
    }]
}
````

Also: needed to rename the `Read` action to `Show` so that Goa
can generate the `Href` utility function to provide a link to the
resource in the JSON-API response.

Fixes #629

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>