Delete user on Che (#813)
Fixes #ODC140

Call Che with a token generated in behalf of the user (see below). Expect a `204` response if all went well, but ignore the error if Che responds with `404`.

Also: use a transient token to call Che. This token has a very short life duration and is not even registered in the database. The validation process skips the database check as it sees the `transient` claim in the token.

Co-authored-by: Shane Bryzak <sbryzak@gmail.com>
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>